+++ lib/hexapdf/type.rb
@@ -61,6 +61,8 @@ module HexaPDF
     autoload(:FontSimple, 'hexapdf/type/font_simple')
     autoload(:FontType1, 'hexapdf/type/font_type1')
     autoload(:FontTrueType, 'hexapdf/type/font_true_type')
+    autoload(:FontType0, 'hexapdf/type/font_type0')
+    autoload(:CIDFont, 'hexapdf/type/cid_font')
 
   end