+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Added
 
+* [HexaPDF::Type::FontType0] and [HexaPDF::Type::CIDFont] for composite font
+  support
 * Complete support for CMaps for use with composite fonts; the interface for
   [HexaPDF::Font::CMap] changed to accomodate this
 * Support for marked-content points and sequences in [HexaPDF::Content::Canvas]